pyproject.toml
[tool.poetry]
name = "django-auth"
version = "0.1.0"
description = ""
authors = ["Eason_chiang <95297596+class83108@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
psycopg2-binary = "^2.9.9"
load-dotenv = "^0.1.0"
django-allauth = {extras = ["socialaccount"], version = "^65.0.0"}
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
